requirements.txt
fastapi
uvicorn
websockets
discord.py
PyNaCl
SQLAlchemy
jinja2
python-multipart
